requirements.txt
djangorestframework==3.13.1
djongo==1.3.6
django-cors-headers==3.13.0
pymongo==3.12.1
